refactor: 통계 로직 훅으로 분리
